added books blueprint
