+++ app.py
@@ -17,6 +17,7 @@ from init import db, ma, bcrypt, jwt
 
 from blueprints.cli_bp import cli_bp
 from blueprints.auth_bp import auth_bp
+from blueprints.books_bp import books_bp
 
 app = Flask(__name__)
 
@@ -30,23 +31,9 @@ bcrypt.init_app(app)
 
 app.register_blueprint(cli_bp)
 app.register_blueprint(auth_bp)
+app.register_blueprint(books_bp)
     
-@app.route("/books", methods=["GET"])
-def books():
-    books_list = Book.query.all()
-    result = books_schema.dump(books_list)
-    return jsonify(result)
 
-
-
-@app.route('/book_details/<int:book_id>', methods=['GET'])
-def book_details(book_id: int):
-    book = Book.query.filter_by(book_id=book_id).first()
-    if book:
-        result = book_schema.dump(book)
-        return jsonify(result)
-    else:
-        return jsonify(message="That book does not exist"), 404
     
 @app.route('/add_review', methods=['POST'])
 @jwt_required()
@@ -155,51 +142,6 @@ def update_review():
     else:
         return jsonify(message="Review does not exist"), 404
 
-@app.route('/add_book', methods=['POST'])
-@jwt_required()
-def add_book():
-    title = request.form['title']
-    test = Book.query.filter_by(title=title).first()
-    if test:
-        return jsonify("There is already a book with that title"), 409
-    else:
-        genre = request.form['genre']
-        author = request.form['author']
-        synopsis = request.form['synopsis']
-        publication_year = int(request.form['publication_year'])
-
-        new_book = Book(title=title, author=author, genre=genre, synopsis=synopsis, publication_year=publication_year)
-
-        db.session.add(new_book)
-        db.session.commit()
-        return jsonify(message="You added a new book!"), 201
-
-@app.route('/update_book', methods=['PUT'])
-@jwt_required()
-def update_book():
-    book_id = int(request.form['book_id'])
-    book = Book.query.filter_by(book_id=book_id).first()
-    if book:
-        book.title = request.form['title']
-        book.author = request.form['author']
-        book.genre = request.form['genre']
-        book.synopsis = request.form['synopsis']
-        book.publication_year = int(request.form['publication_year'])
-        db.session.commit()
-        return jsonify(message="You updated a book!"), 202
-    else:
-        return jsonify(message="Book does not exist"), 404
-
-@app.route('/delete_book/<int:book_id>', methods=['DELETE'])
-@jwt_required()
-def delete_book(book_id: int):
-    book = Book.query.filter_by(book_id=book_id).first()
-    if book:
-        db.session.delete(book)
-        db.session.commit()
-        return jsonify(message="You deleted a book"), 202
-    else: 
-         return jsonify(message="That book does not exist"), 202
     
 @app.route('/delete_review/<int:review_id>', methods=['DELETE'])
 @jwt_required()
